Refactor APIClient to use unified async client
